# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Biblioteca",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/usuarios/{idLibro}": {
            "get": {
                "tags": [
                    "Libros"
                ],
                "summary": "Mostrar un libro por su id",
                "operationId": "25fb622426857416567b80c11fc9ed45",
                "parameters": [
                    {
                        "name": "idLibro",
                        "in": "path",
                        "description": "La id del libro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar un libro por su id."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "patch": {
                "tags": [
                    "Libros"
                ],
                "summary": "Editar Libro",
                "operationId": "31a4c5a8dd8bd1e9c12d63b06fb3e847",
                "parameters": [
                    {
                        "name": "idLibro",
                        "in": "path",
                        "description": "La id del libro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "titulo",
                        "in": "query",
                        "description": "Nuevo titulo de libro",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "descripcion",
                        "in": "query",
                        "description": "Nueva descripcion del libro",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Modificar un libro"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/usuarios/{idlibro} ": {
            "delete": {
                "tags": [
                    "Libros"
                ],
                "summary": "Eliminar Libro",
                "operationId": "6cc8870161c83ae57db742e35f1f3ef6",
                "parameters": [
                    {
                        "name": "idLibro",
                        "in": "path",
                        "description": "La id del libro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Eliminar un libro por su Id y te muestra el libro borrado"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
        "/api/usuarios": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Mostrar todos los usuarios",
                "operationId": "94dc27f677a70fac55451270f4b50765",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Añadir Usuario",
                "operationId": "5ca59aa4a769ef25d650b9affdba01f8",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "query",
                        "description": "Nombre de usuario nuevo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "correo de usuario nuevo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Contraseña de usuario nuevo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "Confirmacion de Contraseña de usuario nuevo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Añadir Usuario y te devuelve el usuario insertado."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/usuarios/{idusuario}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Mostrar un usuario por su id",
                "operationId": "49b788d6df98082222c8d8ab10537a98",
                "parameters": [
                    {
                        "name": "idusuario",
                        "in": "path",
                        "description": "La id del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar un usuario por su id."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "patch": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Editar Usuario",
                "operationId": "4571026da8144edbc2cb4864b1352595",
                "parameters": [
                    {
                        "name": "idusuario",
                        "in": "path",
                        "description": "La id del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "nombre",
                        "in": "query",
                        "description": "Nuevo nombre de usuario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Nuevo email de usuario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Modificar un usuario"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/usuarios/{idusuario} ": {
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Eliminar Usuario",
                "operationId": "8981f2aca6f9889a32864dbc3d5c5920",
                "parameters": [
                    {
                        "name": "idusuario",
                        "in": "path",
                        "description": "La id del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Eliminar un usuario por su Id y te muestra el usuario borrado"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
    }
}